requirements.txt
PyYAML==6.0.2
PyAutoGUI==0.9.54
MouseInfo==0.1.3
PyMsgBox==2.0.1
pillow==12.0.0
pytesseract==0.3.13
mss==10.1.0
numpy==2.2.6
discord.py==2.6.4
requests==2.32.5
certifi==2025.11.12
urllib3==2.6.2
